ModelExporter実装

+++ src/main/java/com/chikage/mineexporter/utils/Vertex.java
@@ -5,7 +5,7 @@ public class Vertex {
     public final float y;
     public final float z;
 
-    public Vertex(int x, int y, int z) {
+    public Vertex(float x, float y, float z) {
         this.x = x;
         this.y = y;
         this.z = z;
